WT_MIN of a uint32_t value has to be cast before being used as an int.

+++ src/reconcile/rec_track.c
@@ -182,7 +182,8 @@ __ovfl_reuse_verbose(WT_SESSION_IMPL *session,
 	    F_ISSET(reuse, WT_OVFL_REUSE_INUSE) &&
 	    F_ISSET(reuse, WT_OVFL_REUSE_JUST_ADDED) ? ", " : "",
 	    F_ISSET(reuse, WT_OVFL_REUSE_JUST_ADDED) ? "just-added" : "",
-	    WT_MIN(reuse->value_size, 40), (char *)WT_OVFL_REUSE_VALUE(reuse));
+	    (int)WT_MIN(reuse->value_size, 40),
+	    (char *)WT_OVFL_REUSE_VALUE(reuse));
 
 	__wt_scr_free(session, &tmp);
 	return (0);
@@ -576,7 +577,8 @@ __ovfl_txnc_verbose(WT_SESSION_IMPL *session,
 	    __wt_addr_string(
 		session, WT_OVFL_TXNC_ADDR(txnc), txnc->addr_size, tmp),
 	    txnc->current,
-	    WT_MIN(txnc->value_size, 40), (char *)WT_OVFL_TXNC_VALUE(txnc));
+	    (int)WT_MIN(txnc->value_size, 40),
+	    (char *)WT_OVFL_TXNC_VALUE(txnc));
 
 	__wt_scr_free(session, &tmp);
 	return (0);
